rename common to interfaces
