+++ service/history/api/respondworkflowtaskcompleted/api_test.go
@@ -62,9 +62,9 @@ import (
 	"go.temporal.io/server/common/testing/updateutils"
 	"go.temporal.io/server/internal/effect"
 	"go.temporal.io/server/service/history/api"
-	history "go.temporal.io/server/service/history/common"
 	"go.temporal.io/server/service/history/events"
 	"go.temporal.io/server/service/history/hsm"
+	history "go.temporal.io/server/service/history/interfaces"
 	"go.temporal.io/server/service/history/shard"
 	"go.temporal.io/server/service/history/tests"
 	"go.temporal.io/server/service/history/workflow"
