+++ service/history/replication/executable_backfill_history_events_task.go
@@ -39,8 +39,8 @@ import (
 	"go.temporal.io/server/common/persistence/versionhistory"
 	serviceerrors "go.temporal.io/server/common/serviceerror"
 	ctasks "go.temporal.io/server/common/tasks"
-	history "go.temporal.io/server/service/history/common"
 	"go.temporal.io/server/service/history/consts"
+	history "go.temporal.io/server/service/history/interfaces"
 )
 
 type (
