+++ service/history/replication/executable_sync_hsm_task.go
@@ -39,7 +39,7 @@ import (
 	serviceerrors "go.temporal.io/server/common/serviceerror"
 	ctasks "go.temporal.io/server/common/tasks"
 	"go.temporal.io/server/service/history/consts"
-	history "go.temporal.io/server/service/history/common"
+	history "go.temporal.io/server/service/history/interfaces"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
