+++ service/history/replication/executable_task_test.go
@@ -59,9 +59,9 @@ import (
 	ctasks "go.temporal.io/server/common/tasks"
 	"go.temporal.io/server/common/testing/protomock"
 	"go.temporal.io/server/common/xdc"
-	history "go.temporal.io/server/service/history/common"
 	"go.temporal.io/server/service/history/configs"
 	"go.temporal.io/server/service/history/consts"
+	history "go.temporal.io/server/service/history/interfaces"
 	"go.temporal.io/server/service/history/replication/eventhandler"
 	"go.temporal.io/server/service/history/shard"
 	"go.temporal.io/server/service/history/tasks"
