+++ service/history/replication/stream_sender_test.go
@@ -45,8 +45,8 @@ import (
 	"go.temporal.io/server/common/namespace"
 	"go.temporal.io/server/common/persistence"
 	serviceerrors "go.temporal.io/server/common/serviceerror"
-	history "go.temporal.io/server/service/history/common"
 	"go.temporal.io/server/service/history/configs"
+	history "go.temporal.io/server/service/history/interfaces"
 	"go.temporal.io/server/service/history/shard"
 	"go.temporal.io/server/service/history/tasks"
 	"go.temporal.io/server/service/history/tests"
