+++ service/history/shard/engine_factory.go
@@ -26,7 +26,7 @@
 
 package shard
 
-import history "go.temporal.io/server/service/history/common"
+import history "go.temporal.io/server/service/history/interfaces"
 
 type (
 	// EngineFactory is used to create an instance of sharded history engine
